@@ -3,36 +3,35 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.7.1",
-    "@emotion/styled": "^11.6.0",
+    "@emotion/react": "^11.8.1",
+    "@emotion/styled": "^11.8.1",
     "@react-icons/all-files": "^4.1.0",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "@types/jest": "^26.0.24",
-    "@types/node": "^12.20.33",
-    "@types/react": "^17.0.30",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-router-dom": "^5.3.1",
-    "axios": "^0.25.0",
-    "framer-motion": "^5.6.0",
-    "i18next": "^21.6.11",
+    "@testing-library/jest-dom": "^5.16.2",
+    "@testing-library/react": "^12.1.3",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^27.4.1",
+    "@types/node": "^17.0.21",
+    "@types/react": "^17.0.39",
+    "@types/react-dom": "^17.0.11",
+    "@types/react-router-dom": "^5.3.3",
+    "axios": "^0.26.0",
+    "framer-motion": "^6.2.8",
+    "i18next": "^21.6.12",
     "i18next-browser-languagedetector": "^6.1.3",
-    "i18next-xhr-backend": "^3.2.2",
     "leaflet": "^1.7.1",
-    "node-sass": "^6.0.1",
+    "node-sass": "^7.0.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-i18next": "^11.15.4",
+    "react-i18next": "^11.15.5",
     "react-icons": "^4.3.1",
-    "react-leaflet": "^3.2.2",
-    "react-router": "^5.2.1",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "4.0.3",
-    "sass": "^1.43.4",
+    "react-leaflet": "^3.2.5",
+    "react-router": "^6.2.1",
+    "react-router-dom": "^6.2.1",
+    "react-scripts": "^5.0.0",
+    "sass": "^1.49.9",
     "styled-components": "^5.3.3",
-    "typescript": "^4.4.4",
-    "web-vitals": "^1.1.2"
+    "typescript": "^4.5.5",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "test": "echo \"No tests specified\" && exit 0",
@@ -56,7 +55,7 @@
     "not op_mini all"
   ],
   "devDependencies": {
-    "@types/leaflet": "^1.7.5",
-    "@types/styled-components": "^5.1.15"
+    "@types/leaflet": "^1.7.9",
+    "@types/styled-components": "^5.1.23"
   }
 }
