@@ -7,27 +7,26 @@ import Map from './components/Map/Map';
 import Imprint from './pages/imprint';
 import Header from './components/Header';
 import Footer from './components/Footer';
-import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
+import { BrowserRouter, Routes, Route} from 'react-router-dom';
 import Navbar from './components/Navbar/Navbar'
 
 
 
 function App() {
   return (
-    <Router basename={`/${process.env.PUBLIC_URL}`}>
-      <Header />
-      {/*<Navigation />*/}
-      <Navbar />
+    <BrowserRouter basename={`/${process.env.PUBLIC_URL}`}>
+        <Header />
+        <Navbar />
       <div id="container">
-      <Switch>
-        <Route path='/' exact component={Home} />
-        <Route path='/about' component={About} />
-        <Route path='/map' component={Map} />
-        <Route path='/imprint' component={Imprint} />
-      </Switch>
+      <Routes>
+        <Route path='/' element={<Home />} />
+        <Route path='about' element={<About />} />
+        <Route path='map' element={<Map />} />
+        <Route path='imprint' element={<Imprint />} />
+        </Routes>
       </div>
       <Footer />
-    </Router>
+      </BrowserRouter>
   );
 }
 
